Сode optimization
Moved the controllers and Joi schema to separate files.
