+++ routes/api/contacts.js
@@ -1,102 +1,19 @@
 const express = require("express");
 const router = express.Router();
-const createError = require("http-errors");
-const Joi = require("joi");
 
-const contactSchema = Joi.object({
-  name: Joi.string().required(),
-  email: Joi.string()
-    .email({ minDomainSegments: 2, tlds: { allow: ["com", "net"] } })
-    .required(),
-  phone: Joi.alternatives([Joi.string(), Joi.number()]).required(),
-});
+const { validation } = require("../../middlewares");
+const contactSchema = require("../../schemas/contacts");
 
-const contactsOperations = require("../../models/contacts");
+const ctrl = require("../../controllers/contacts");
 
-router.get("/", async (req, res, next) => {
-  try {
-    const contacts = await contactsOperations.listContacts();
-    res.json({
-      status: "success",
-      code: 200,
-      data: { result: contacts },
-    });
-  } catch (error) {
-    next(error);
-  }
-});
+router.get("/", ctrl.getAll);
 
-router.get("/:contactId", async (req, res, next) => {
-  try {
-    const { contactId } = req.params;
-    const result = await contactsOperations.getContactById(contactId);
-    if (!result) {
-      throw createError(404, `Contact with id-${contactId} not found`);
-    }
-    res.json({ status: "success", code: 200, data: { result } });
-  } catch (error) {
-    next(error);
-  }
-});
+router.get("/:contactId", ctrl.getById);
 
-router.post("/", async (req, res, next) => {
-  try {
-    const { error } = contactSchema.validate(req.body);
-    if (error) {
-      error.status = 400;
-      error.message = "missing required name field";
-      throw error;
-    }
-    const result = await contactsOperations.addContact(req.body);
-    res.status(201).json({
-      status: "success",
-      code: 201,
-      data: { result },
-    });
-  } catch (error) {
-    next(error);
-  }
-});
+router.post("/", validation(contactSchema), ctrl.add);
 
-router.delete("/:contactId", async (req, res, next) => {
-  try {
-    const { contactId } = req.params;
-    const result = await contactsOperations.removeContact(contactId);
-    if (!result) {
-      throw createError(404, `Contact with id-${contactId} not found`);
-    }
-    res.json({
-      status: "success",
-      code: 200,
-      message: "contact deleted",
-      data: { result },
-    });
-  } catch (error) {
-    next(error);
-  }
-});
+router.delete("/:contactId", ctrl.deleteById);
 
-router.put("/:contactId", async (req, res, next) => {
-  try {
-    const { error } = contactSchema.validate(req.body);
-    if (error) {
-      error.status = 400;
-      error.message = "missing fields";
-      throw error;
-    }
-    const { contactId } = req.params;
-    const result = await contactsOperations.updateContact(contactId, req.body);
-    if (!result) {
-      throw createError(404, `Contact with id-${contactId} not found`);
-    }
-    res.json({
-      status: "success",
-      code: 200,
-      data: { result },
-    });
-  } catch (error) {
-    next(error);
-  }
-});
+router.put("/:contactId", validation(contactSchema), ctrl.updateById);
 
 module.exports = router;
