+++ schemas/contacts.js
@@ -1,0 +1,11 @@
+const Joi = require("joi");
+
+const contactSchema = Joi.object({
+  name: Joi.string().required(),
+  email: Joi.string()
+    .email({ minDomainSegments: 2, tlds: { allow: ["com", "net"] } })
+    .required(),
+  phone: Joi.alternatives([Joi.string(), Joi.number()]).required(),
+});
+
+module.exports = contactSchema;
